# src/AlgoJudge.Application/Contracts/Auth/AuthResponse.cs
namespace AlgoJudge.Application.Contracts.Auth;

public sealed class AuthResponse
{
    public string UserName { get; init; } = string.Empty;
    public string Email { get; init; } = string.Empty;
    public bool IsAdmin { get; init; }
    public DateTime ExpiresAt { get; init; }
}